Build edges and nodes for component tree
