+++ assets/src/components/cd/pipelines/nodes/ComponentTreeNode.tsx
@@ -1,42 +1,19 @@
-import {
-  Card,
-  Chip,
-  CloseRoundedIcon,
-  Spinner,
-  StatusOkIcon,
-  Tooltip,
-} from '@pluralsh/design-system'
-import { GateState, PipelineStageEdgeFragment } from 'generated/graphql'
-import {
-  ComponentProps,
-  ComponentPropsWithoutRef,
-  ReactElement,
-  ReactNode,
-  cloneElement,
-  useMemo,
-} from 'react'
-import {
-  Handle,
-  type Node,
-  type NodeProps,
-  Position,
-  useNodes,
-} from 'reactflow'
+import { Card } from '@pluralsh/design-system'
+import { PipelineStageEdgeFragment } from 'generated/graphql'
+import { ComponentProps, ReactElement, ReactNode, cloneElement } from 'react'
+import { Handle, type NodeProps, Position } from 'reactflow'
 import styled, { useTheme } from 'styled-components'
 import isEmpty from 'lodash/isEmpty'
 
+import { TreeNodeMeta } from 'components/component/ComponentTree'
+
+import ComponentCard from 'components/apps/app/components/ComponentCard'
+
 import { useNodeEdges } from '../utils/hooks'
-import { reduceGateStates } from '../utils/reduceGateStatuses'
 
 export type CardStatus = 'ok' | 'closed' | 'pending'
 
 const HANDLE_SIZE = 10
-
-export const gateStateToCardStatus = {
-  [GateState.Open]: 'ok',
-  [GateState.Closed]: 'closed',
-  [GateState.Pending]: 'pending',
-} as const satisfies Record<GateState, CardStatus>
 
 export const NodeCardList = styled.ul(({ theme }) => ({
   display: 'flex',
@@ -103,28 +80,13 @@ const HandleSC = styled(Handle)<{ $isConnected?: boolean; $isOpen?: boolean }>(
   })
 )
 
-export function BaseNode({
+export function ComponentTreeNode({
   id,
-  data: { meta },
-  children,
+  data,
   ...props
-}: NodeProps<NodeMeta> & { children: ReactNode } & ComponentProps<
-    typeof BaseNodeSC
-  >) {
+}: NodeProps<TreeNodeMeta> & ComponentProps<typeof BaseNodeSC>) {
   const { incomers, outgoers } = useNodeEdges(id)
-  const nodes = useNodes()
-
-  const reducedInState = useMemo(() => {
-    const incomingNodes = nodes.filter((node) =>
-      incomers.some((incomer) => incomer.source === node.id)
-    )
-
-    return reduceGateStates(
-      incomingNodes.map((inNode) => ({
-        state: (inNode as Node<NodeMeta>)?.data?.meta?.state,
-      }))
-    )
-  }, [incomers, nodes])
+  const { metadata } = data
 
   return (
     <BaseNodeSC {...props}>
@@ -132,15 +94,17 @@ export function BaseNode({
         type="target"
         isConnectable={false}
         $isConnected={!isEmpty(incomers)}
-        $isOpen={reducedInState === GateState.Open}
+        $isOpen
         position={Position.Left}
       />
-      {children}
+      <ComponentCard
+        component={{ name: metadata?.name || '', kind: data.type }}
+      />
       <HandleSC
         type="source"
         isConnectable={false}
         $isConnected={!isEmpty(outgoers)}
-        $isOpen={meta.state === GateState.Open}
+        $isOpen
         position={Position.Right}
       />
     </BaseNodeSC>
@@ -153,12 +117,6 @@ const IconHeadingSC = styled.div(({ theme }) => ({
   gap: theme.spacing.xsmall,
   ...theme.partials.text.body2Bold,
 }))
-
-export const gateStateToSeverity = {
-  [GateState.Open]: 'success',
-  [GateState.Closed]: 'critical',
-  [GateState.Pending]: 'warning',
-} as const satisfies Record<GateState, ComponentProps<typeof Chip>['severity']>
 
 export function IconHeading({
   icon,
@@ -181,62 +139,4 @@ export function IconHeading({
   )
 }
 
-export type NodeMeta = { meta: { state: GateState } }
-
-export type EdgeNode = NodeProps<PipelineStageEdgeFragment & NodeMeta>
-
-const StatusCardSC = styled(Card)(({ theme }) => ({
-  '&&': {
-    padding: `${theme.spacing.xsmall}px ${theme.spacing.small}px`,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    gap: theme.spacing.small,
-  },
-  '.state': {
-    display: 'flex',
-    alignItems: 'center',
-  },
-}))
-
-export function StatusCard({
-  status,
-  statusLabel,
-  children,
-  ...props
-}: {
-  status: Nullable<CardStatus>
-  statusLabel?: Nullable<string>
-} & ComponentPropsWithoutRef<typeof StatusCardSC>) {
-  const theme = useTheme()
-
-  return (
-    <StatusCardSC {...props}>
-      <div className="contentArea">{children}</div>
-      {status && (
-        <div className="state">
-          <Tooltip label={statusLabel}>
-            {status === 'ok' ? (
-              <StatusOkIcon
-                size={12}
-                color={theme.colors['icon-success']}
-              />
-            ) : status === 'closed' ? (
-              <CloseRoundedIcon
-                size={12}
-                color={theme.colors['icon-danger-critical']}
-              />
-            ) : (
-              <div>
-                <Spinner
-                  size={12}
-                  color={theme.colors['icon-warning']}
-                />
-              </div>
-            )}
-          </Tooltip>
-        </div>
-      )}
-    </StatusCardSC>
-  )
-}
+export type EdgeNode = NodeProps<PipelineStageEdgeFragment & TreeNodeMeta>
