+++ assets/src/graph/kubernetes/kubernetes.graphql
@@ -1,4 +1,5 @@
 fragment Metadata on Metadata {
+  uid
   name
   namespace
   labels {
